# openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: XpressPro FX Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trades/{tradeId}/release:
    post:
      operationId: releaseTradeFunds
      tags: [trades]
      summary: Release completed trade funds to main wallet
      parameters:
        - name: tradeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Release result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
...
  /messages:
    get:
      operationId: getMessages
      tags: [messages]
      summary: Get messages with manager or in P2P context
      parameters:
        - name: context
          in: query
          required: false
          schema:
            type: string
            enum: [manager, p2p, support]
        - name: contextId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "200":
          description: Sent message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
  /mailbox:
    get:
      operationId: getMailbox
      tags: [support]
      summary: Get the user's mailbox threads
      responses:
        "200":
          description: Mailbox threads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MailboxThread"
    post:
      operationId: sendMailboxMessage
      tags: [support]
      summary: Send an email to a platform address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMailBody"
      responses:
        "200":
          description: Created thread
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MailboxThread"
  /mailbox/{threadId}/reply:
    post:
      operationId: userMailboxReply
      tags: [support]
      summary: User replies to one of their existing mailbox threads (blocked when thread.noReply is true)
      parameters:
        - name: threadId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MailboxUserReplyBody"
      responses:
        "200":
          description: Updated thread
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MailboxThread"
...
  /admin/mailbox/{threadId}/reply:
    post:
      operationId: adminMailboxReply
      tags: [admin]
      summary: Admin replies to a mailbox thread
      parameters:
        - name: threadId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MailboxReplyBody"
      responses:
        "200":
          description: Updated thread
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MailboxThread"
...
components:
  schemas:
...
    ActionResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
...
    Message:
      type: object
      properties:
        id:
          type: string
        senderId:
          type: string
        senderName:
          type: string
        senderAvatar:
          type: string
          nullable: true
        content:
          type: string
        context:
          type: string
          enum: [manager, p2p, p2p_admin, support]
        contextId:
          type: string
          nullable: true
        isFromUser:
          type: boolean
        createdAt:
          type: string
      required: [id, senderId, senderName, content, context, isFromUser, createdAt]
    SendMessageRequest:
      type: object
      properties:
        content:
          type: string
        context:
          type: string
          enum: [manager, p2p, p2p_admin, support]
        contextId:
          type: string
          nullable: true
      required: [content, context]
...
    MailboxMessage:
      type: object
      properties:
        id:
          type: string
        from:
          type: string
        content:
          type: string
        imageUrl:
          type: string
          nullable: true
          description: Optional inline image attachment URL.
        createdAt:
          type: string
      required: [id, from, content, createdAt]
    MailboxThread:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        from:
          type: string
        to:
          type: string
        subject:
          type: string
        messages:
          type: array
          items:
            $ref: "#/components/schemas/MailboxMessage"
        read:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
        noReply:
          type: boolean
          description: When true, the user cannot reply to this thread (admin-controlled per-thread no-reply enforcement).
      required: [id, userId, from, to, subject, messages, read, createdAt, updatedAt]
    SendMailBody:
      type: object
      properties:
        to:
          type: string
          description: "Recipient mailbox (e.g. help@xpressprofx.com)"
        subject:
          type: string
        content:
          type: string
        imageUrl:
          type: string
          nullable: true
      required: [to, subject, content]
    MailboxReplyBody:
      type: object
      properties:
        from:
          type: string
          description: Which platform address is replying
        content:
          type: string
        imageUrl:
          type: string
          nullable: true
        noReply:
          type: boolean
          description: When true, lock this thread so the user cannot reply to it.
      required: [from, content]
    MailboxUserReplyBody:
      type: object
      properties:
        content:
          type: string
        imageUrl:
          type: string
          nullable: true
      required: [content]
...